requirements.txt
streamlit
tensorflow
pillow
opencv-python-headless
